requirements.txt
# main speech recognition module
openai-whisper

# webvtt subtitle module
webvtt-py

# grammar check
language_tool_python